chore: 試合のヘッダーを切り出した
